added working MVP
Added the working MVP of the application, which consists of a single screen that shows
the instant availability/unavailability of all ISEP rooms.
This MVP works from start to end and fetches data from our own web API.
The data is then cached locally on a SQLite database.
The screen can be refreshed by the user to force re-querying the API and rebuilding the local database.
Both light and dark themes have been overriden, which means that the app works correctly both in
light and dark themes.
Room availability is calculated locally (on the phone) in real-time.

TODO: Translations